requirements.txt
fastapi
uvicorn[standard]
mt5linux
pandas
python-dotenv
